create_rental class and person_choice class
